Fix: исправлены замечания detect, вторая итерация
