+++ app/src/main/java/ru/practicum/android/diploma/team/information/presentation/ui/TeanInformationScreen.kt
@@ -1,4 +1,4 @@
-package ru.practicum.android.diploma.team_information.presentation.ui
+package ru.practicum.android.diploma.team.information.presentation.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
